% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorapp.R
\name{explorapp}
\alias{explorapp}
\title{Shiny app to compare PPtree, PPtreeExt and rpart boundaries in 2D with different simulation scenarios}
\usage{
explorapp(ui, server)
}
\arguments{
\item{ui}{user interface}

\item{server}{server function}
}
\description{
Shiny app to compare PPtree, PPtreeExt and rpart boundaries in 2D with different simulation scenarios
}
\examples{
if(interactive()){
explorapp(ui,server)
}

}
