% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findproj_MOD.R
\name{findproj_MOD}
\alias{findproj_MOD}
\title{1D projection for each node partition using entropy}
\usage{
findproj_MOD(origclass, origdata, PPmethod = "LDA", q = 1, weight = TRUE, lambda = .1)
}
\arguments{
\item{origclass}{original class}

\item{origdata}{original data}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA"}

\item{q}{1D proj}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{lambda}{lambda in PDA index}
}
\description{
1D projection for each node partition using entropy
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection Pursuit Classification Tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
