% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAopt_MOD.R
\name{LDAopt_MOD}
\alias{LDAopt_MOD}
\title{PP optimization using LDA index}
\usage{
LDAopt_MOD(origclass,origdata,q=1,weight=TRUE,...)
}
\arguments{
\item{origclass}{class information vector of data}

\item{origdata}{data matrix without class information}

\item{q}{dimension of projection vector}

\item{weight}{weight flag in LDA index}

\item{...}{arguments to be passed to methods}
}
\value{
indexbest maximum LDA index value

projbest optimal q-dimensional projection matrix

origclass original class information vector

origdata  original data matrix  without class information
}
\description{
PP optimization using LDA index same as original
}
\details{
Find the q-dimensional optimal projection using LDA projectin pursuit index
}
\references{
Lee, EK., Cook, D., Klinke, S., and Lumley, T.(2005) 
Projection Pursuit for Exploratory Supervised Classification, 
Journal of Computational and Graphical Statistics, 14(4):831-846.
}
