% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPtreeExt_split.R
\name{PPtreeExt_split}
\alias{PPtreeExt_split}
\title{Projection pursuit classification tree with random variable selection in each split}
\usage{
PPtreeExt_split(formula, data, PPmethod = 'LDA', 
size.p = 1,  lambda = 0.1, entro = FALSE, entroindiv = FALSE,...)
}
\arguments{
\item{formula}{A character with the name of the class variable.}

\item{data}{Data frame with the complete data set.}

\item{PPmethod}{index to use for projection pursuit: 'LDA', 'PDA'}

\item{size.p}{proportion of variables randomly sampled in each split, default is 1, returns a PPtree.}

\item{lambda}{penalty parameter in PDA index and is between 0 to 1 . If \code{lambda = 0}, no penalty parameter is added and the PDA index is the same as LDA index. If \code{lambda = 1} all variables are treated as uncorrelated. The default value is \code{lambda = 0.1}.}

\item{entro}{FALSE, compute the entropy method}

\item{entroindiv}{FALSE, compute the entropy for each observation in 1D projection}

\item{...}{arguments to be passed to methods}
}
\value{
An object of class \code{PPtreeclass} with components
\item{Tree.Struct}{Tree structure of projection pursuit classification tree}
\item{projbest.node}{1-dim optimal projections of each split node}
\item{splitCutoff.node}{cutoff values of each split node}
\item{origclass}{original class} 
\item{origdata}{original data}
}
\description{
Find tree structure using various projection pursuit indices of classification in each split.
}
\examples{
data(penguins)
penguins <- na.omit(penguins[, -c(2,7, 8)])
require(rsample)
penguins_spl <- rsample::initial_split(penguins, strata=species)
penguins_train <- training(penguins_spl)
penguins_test <- testing(penguins_spl)
penguins_ppt2 <- PPtreeExt_split(species~bill_len + bill_dep +
flipper_len + body_mass, data = penguins_train, PPmethod = "LDA", tot=nrow
(penguins_train), tol =  0.5 , entro=TRUE)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK (2013) 
PPtree: Projection pursuit classification tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}
