% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPtreeExtclass.R
\name{PPtreeExtclass}
\alias{PPtreeExtclass}
\title{Projection pursuit classification tree}
\usage{
PPtreeExtclass(formula,data, PPmethod = "LDA",weight = TRUE,
                     r = 1,lambda = 0.1, energy = 0,maxiter = 50000, strule = 1,tot,...)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA"}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{r}{r in Lr index}

\item{lambda}{lambda in PDA index}

\item{energy}{parameter for the probability to take new projection}

\item{maxiter}{maximum iteration number}

\item{strule}{select the stoping rule, 1 all observations in the node belongs 
to the same class based, 2 node size is less than 5% of the total number of observations, 
3 the entropy reduction is samaller than a treshold.}

\item{tot}{total obs original class}

\item{...}{arguments to be passed to methods}
}
\value{
Tree.Struct tree structure of projection pursuit classification tree

projbest.node 1 dimensional optimal projections of each node split

splitCutoff.node cutoff values of each node split

origclass original class

origdata original data
}
\description{
Construct the projection pursuit classification tree extensions
Find tree structure using various projection pursuit indices of 
classification in each split.
}
\examples{
data(penguins)
penguins <- na.omit(penguins[, -c(2,7)])
penguins_ppt <- PPtreeExtclass(species~bill_len + bill_dep +
  flipper_len + body_mass, data = penguins, PPmethod = "PDA")
penguins_ppt
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection Pursuit Classification Tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}
