% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.PPtreeclassMOD.R
\name{predict.PPtreeclassMOD}
\alias{predict.PPtreeclassMOD}
\title{predict PPtreeExt}
\usage{
\method{predict}{PPtreeclassMOD}(object, newdata, true.class = NULL, ...)
}
\arguments{
\item{object}{PPtreeclass object}

\item{newdata}{the test dataset}

\item{true.class}{true class of test dataset if available}

\item{...}{arguments to be passed to methods}
}
\value{
A list with:
\describe{
  \item{predict.class }{predicted class}
  \item{predict.error}{number of the prediction errors}
}
}
\description{
Predict Predict method for projection pursuit 
classification tree Extension and calculate prediction error.
}
\examples{
data(penguins)
penguins <- na.omit(penguins[, -c(2,7, 8)])
require(rsample)
penguins_spl <- rsample::initial_split(penguins, strata=species)
penguins_train <- training(penguins_spl)
penguins_test <- testing(penguins_spl)
penguins_ppt <- PPTreeclass_MOD(species~bill_len + bill_dep +
  flipper_len + body_mass, data = penguins_train, PPmethod = "LDA")
predict(object = penguins_ppt, newdata =penguins_test[,-1], true.class = penguins_test$species)
}
\keyword{tree}
