% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PPtreeclassMOD.R
\name{print.PPtreeclassMOD}
\alias{print.PPtreeclassMOD}
\title{Print PPtreeclassMOD object}
\usage{
\method{print}{PPtreeclassMOD}(x, coef.print = FALSE, cutoff.print = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{something}

\item{coef.print}{something}

\item{cutoff.print}{something}

\item{verbose}{something}

\item{...}{something}
}
\description{
Print PPtreeclassMOD object
}
