% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PPtreeExtclass.R
\name{print.PPtreeExtclass}
\alias{print.PPtreeExtclass}
\title{Print PPtreeExtclass object}
\usage{
\method{print}{PPtreeExtclass}(x, coef.print = FALSE, cutoff.print = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{something}

\item{coef.print}{something}

\item{cutoff.print}{something}

\item{verbose}{something}

\item{...}{something}
}
\description{
Print PPtreeExtclass object
}
