% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree.construct_MOD.R
\name{Tree.construct_MOD}
\alias{Tree.construct_MOD}
\title{Projection pursuit classification tree MOD}
\usage{
Tree.construct_MOD(origclass, origdata, Tree.Struct, id, rep, rep1, rep2,
projbest.node, splitCutoff.node, PPmethod, r = NULL, 
lambda = NULL, TOL, maxiter = 50000, q = 1, weight = TRUE, tol = .5, strule, tot,...)
}
\arguments{
\item{origclass}{original class}

\item{origdata}{original data}

\item{Tree.Struct}{tree structure of projection pursuit classification tree}

\item{id}{identify}

\item{rep}{something}

\item{rep1}{something}

\item{rep2}{something}

\item{projbest.node}{something}

\item{splitCutoff.node}{something}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA"}

\item{r}{r in Lr index}

\item{lambda}{weight flag in PDA index}

\item{TOL}{tolerance}

\item{maxiter}{something}

\item{q}{1D proj}

\item{weight}{weight flag in LDA, PDA}

\item{tol}{tolerance}

\item{strule}{select the stoping rule based, 1 = pure node, 2 = size of node, 3 = entropy}

\item{tot}{total observarions of the original class}

\item{...}{arguments to be passed to methods}
}
\description{
Construct the projection pursuit classification tree MOD
}
\details{
Find tree structure using various projection pursuit indices of 
classification in each split.
}
