% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorapp.R
\name{explorapp}
\alias{explorapp}
\title{Shiny app to explore PPtree algirith partitions}
\usage{
explorapp(ui, server)
}
\arguments{
\item{ui}{A}

\item{server}{aaa}
}
\description{
Shiny app to explore PPtree algirith partitions
}
\examples{
if(interactive()){
explorapp(ui,server)
}
}
