% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findproj_Ext.R
\name{findproj_Ext}
\alias{findproj_Ext}
\title{1D projection for each node partition using entropy}
\usage{
findproj_Ext(origclass, origdata, PPmethod = "LDA", q = 1, weight = TRUE, lambda = .1)
}
\arguments{
\item{origclass}{original class}

\item{origdata}{original data}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA"}

\item{q}{numeric value with dimension of the projected data, if it is 1 then 1D projection is used}

\item{weight}{weight flag in LDA, PDA or index}

\item{lambda}{lambda in PDA index}
}
\description{
1D projection for each node partition using entropy
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK (2013) 
PPtree: Projection Pursuit Classification Tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
