% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDAopt_Ext.R
\name{PDAopt_Ext}
\alias{PDAopt_Ext}
\title{PP optimization using PDA index}
\usage{
PDAopt_Ext(origclass,origdata, q = 1, weight = TRUE, lambda = 0.1,...)
}
\arguments{
\item{origclass}{class information vector of data}

\item{origdata}{data matrix without class information}

\item{q}{dimension of projection vector}

\item{weight}{weight flag in PDA index}

\item{lambda}{lambda in PDA index}

\item{...}{arguments to be passed to methods}
}
\value{
indexbest maximum PDA index value

projbest optimal q-dimensional projection matrix

origclass original class information vector

origdata  original data matrix without class information
}
\description{
PP optimization using PDA index same as PPtree
}
\details{
Find the q-dimensional optimal projection using PDA projectin pursuit index
}
\examples{
data(penguins)
penguins <- na.omit(penguins[, -c(2,7)])
penguins_pda_proj <- PDAopt_Ext(penguins[,1],penguins[,-1], weight=TRUE, q=2, lambda=0.1)
penguins_pda_proj$indexbest
penguins_pda_proj$projbest
}
\references{
Lee, EK, Cook, D.(2010) 
A Projection Pursuit Index for Large p Small n Data, 
Statistics and Computing, 20:381-392.
}
\keyword{projection}
\keyword{pursuit}
