% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPclassify_MOD.R
\name{PPclassify_MOD}
\alias{PPclassify_MOD}
\title{predict PPtree}
\usage{
PPclassify_MOD(Tree.result,test.data,true.class=NULL,...)
}
\arguments{
\item{Tree.result}{PPtreeclass object}

\item{test.data}{the test dataset}

\item{true.class}{true class of test dataset if available}

\item{...}{arguments to be passed to methods}
}
\value{
predict.class predicted class

predict.error number of the prediction errors
}
\description{
predict projection pursuit classification tree
}
\details{
Predict class for the test set with the fitted projection pursuit 
classification tree and calculate prediction error.
}
\examples{
data(iris)
n <- nrow(iris)
tot <- c(1:n)
n.train <- round(n*0.9)
train <- sample(tot,n.train)
test <- tot[-train]
Tree.result <- PPTreeclass_MOD(Species~.,data=iris[train,],"LDA")
PPclassify_MOD(Tree.result,iris[test,1:4],iris[test,5])

}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection Pursuit Classification Tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}
