% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPtreeExtclass.R
\name{PPtreeExtclass}
\alias{PPtreeExtclass}
\title{Projection pursuit classification tree}
\usage{
PPtreeExtclass(formula, data, PPmethod = "LDA", weight = TRUE,
                   lambda = 0.1,srule, tot=nrow(data), tol =0.5,...)
}
\arguments{
\item{formula}{an object of class "formula"}

\item{data}{data frame}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA"}

\item{weight}{weight flag in LDA, PDA and Lr index}

\item{lambda}{lambda in PDA index}

\item{srule}{srule stopping rule flag; if TRUE use stopping rule, if FALSE stop
only for pure or empty nodes}

\item{tot}{total number of observations}

\item{tol}{tolerance value for stopping rule}

\item{...}{arguments to be passed to methods}
}
\value{
Tree.Struct tree structure of projection pursuit classification tree

projbest.node 1 dimensional optimal projections of each node split

splitCutoff.node cutoff values of each node split

origclass original class

origdata original data
}
\description{
Construct the projection pursuit classification tree extensions
Find tree structure using various projection pursuit indices of 
classification in each split.
}
\examples{
data(penguins)
penguins <- na.omit(penguins[, -c(2,7, 8)])
require(rsample)
penguins_spl <- rsample::initial_split(penguins, strata=species)
penguins_train <- training(penguins_spl)
penguins_test <- testing(penguins_spl)
penguins_ppt <- PPtreeExtclass(species~bill_len + bill_dep +
flipper_len + body_mass, data = penguins_train, PPmethod = "LDA", tot=nrow
(penguins_train), tol=0.5 )
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection Pursuit Classification Tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}
