% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree.construct_MOD.R
\name{Tree.construct_MOD}
\alias{Tree.construct_MOD}
\title{Projection pursuit classification tree MOD}
\usage{
Tree.construct_MOD(origclass,origdata,Tree.Struct, id,rep,rep1,rep2,
projbest.node,splitCutoff.node,PPmethod,r = NULL, 
lambda=NULL,TOL,maxiter=50000,q=1,weight=TRUE,tol = .5,...)
}
\arguments{
\item{origclass}{original class}

\item{origdata}{original data}

\item{Tree.Struct}{tree structure of projection pursuit classification tree}

\item{id}{something}

\item{rep}{something}

\item{rep1}{something}

\item{rep2}{something}

\item{projbest.node}{somenthing}

\item{splitCutoff.node}{something}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA"}

\item{r}{r in Lr index}

\item{lambda}{something}

\item{TOL}{dfasd}

\item{maxiter}{something}

\item{q}{something}

\item{weight}{weight flag in LDA, PDA}

\item{tol}{something}

\item{...}{something}
}
\description{
Construct the projection pursuit classification tree MOD (NEW)
}
\details{
Find tree structure using various projection pursuit indices of 
classification in each split.
}
\examples{
data(iris)
Tree.result <- PPtreeViz::PPTreeclass(Species~.,data = iris,"LDA")
Tree.result
}
