% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeExt.construct.R
\name{TreeExt.construct}
\alias{TreeExt.construct}
\title{Projection pursuit classification tree MOD}
\usage{
TreeExt.construct(origclass, origdata, Tree.Struct, id, rep, rep1, rep2,
projbest.node, splitCutoff.node, PPmethod, 
lambda = NULL, q = 1, weight = TRUE, srule=TRUE, tot=NULL, tol = .5,...)
}
\arguments{
\item{origclass}{original class}

\item{origdata}{original data}

\item{Tree.Struct}{tree structure of projection pursuit classification tree}

\item{id}{tree node id}

\item{rep}{internal counter for nodes}

\item{rep1}{internal counter for nodes}

\item{rep2}{internal counter for nodes}

\item{projbest.node}{bests projection node}

\item{splitCutoff.node}{cutof node}

\item{PPmethod}{method for projection pursuit; "LDA", "PDA"}

\item{lambda}{lambda in PDA index}

\item{q}{numeric value with dimension of the projected data, if it is 1 then 1D projection is used}

\item{weight}{weight flag in LDA, PDA}

\item{srule}{stopping rule flag; if TRUE use stopping rule, if FALSE stop only for pure or empty nodes}

\item{tot}{total number of observations}

\item{tol}{tolerance value for entropy stopping rule for splitting a node}

\item{...}{additional arguments to pass trough}
}
\description{
Construct the projection pursuit classification tree with the extensions
}
\details{
Find tree structure using various projection pursuit indices of 
classification in each split.
}
